[REF] mail: move messaging_components, services/, utils/ to legacy/
