+++ addons/mail/static/src/legacy/components/call_invite_request_popup_list/call_invite_request_popup_list.js
@@ -1,6 +1,6 @@
 /** @odoo-module **/
 
-import { registerMessagingComponent } from "@mail/utils/messaging_component";
+import { registerMessagingComponent } from "@mail/legacy/utils/messaging_component";
 
 import { Component } from "@odoo/owl";
