+++ addons/mail/static/src/legacy/components/call_settings_menu_device/call_settings_menu_device.js
@@ -1,6 +1,6 @@
 /** @odoo-module **/
 
-import { registerMessagingComponent } from "@mail/utils/messaging_component";
+import { registerMessagingComponent } from "@mail/legacy/utils/messaging_component";
 
 import { Component } from "@odoo/owl";
