+++ addons/mail/static/src/legacy/components/thread_needaction_preview/thread_needaction_preview.js
@@ -1,7 +1,7 @@
 /** @odoo-module **/
 
 import { useRefToModel } from "@mail/legacy/component_hooks/use_ref_to_model";
-import { registerMessagingComponent } from "@mail/utils/messaging_component";
+import { registerMessagingComponent } from "@mail/legacy/utils/messaging_component";
 
 import { Component } from "@odoo/owl";
