+++ addons/mail/static/src/legacy/components/tracking_value/tracking_value.js
@@ -1,6 +1,6 @@
 /** @odoo-module **/
 
-import { registerMessagingComponent } from "@mail/utils/messaging_component";
+import { registerMessagingComponent } from "@mail/legacy/utils/messaging_component";
 
 import { Component } from "@odoo/owl";
