+++ addons/mail/static/src/legacy/services/messaging_service.js
