+++ addons/mail/static/src/legacy/utils/deferred.js
