+++ addons/mail/static/src/legacy/utils/make_messaging_to_legacy_env.js
