+++ addons/mail/static/src/legacy/utils/media_monitoring.js
