+++ addons/mail/static/src/legacy/utils/messaging_component.js
