+++ addons/mail/static/src/main.js
@@ -2,8 +2,8 @@
 
 import { DialogManagerContainer } from "@mail/legacy/components/dialog_manager_container/dialog_manager_container";
 import { PopoverManagerContainer } from "@mail/legacy/components/popover_manager_container/popover_manager_container";
-import { messagingService } from "@mail/services/messaging_service";
-import { makeMessagingToLegacyEnv } from "@mail/utils/make_messaging_to_legacy_env";
+import { messagingService } from "@mail/legacy/services/messaging_service";
+import { makeMessagingToLegacyEnv } from "@mail/legacy/utils/make_messaging_to_legacy_env";
 
 import { registry } from "@web/core/registry";
