+++ addons/mail/static/tests/new/discuss/sidebar_tests.js
@@ -8,7 +8,7 @@ import {
     start,
     startServer,
 } from "@mail/../tests/helpers/test_utils";
-import { makeDeferred } from "@mail/utils/deferred";
+import { makeDeferred } from "@mail/legacy/utils/deferred";
 import { editInput, getFixture } from "@web/../tests/helpers/utils";
 import { makeFakeNotificationService } from "@web/../tests/helpers/mock_services";
