+++ addons/mail/static/tests/new/thread/thread_tests.js
@@ -12,7 +12,7 @@ import {
     start,
     startServer,
 } from "@mail/../tests/helpers/test_utils";
-import { makeDeferred } from "@mail/utils/deferred";
+import { makeDeferred } from "@mail/legacy/utils/deferred";
 
 import { getFixture, nextTick, patchWithCleanup } from "@web/../tests/helpers/utils";
