+++ addons/note/static/src/legacy/components/activity_menu_view/activity_menu_view.js
@@ -3,7 +3,7 @@
 import { useRefToModel } from '@mail/legacy/component_hooks/use_ref_to_model';
 import { useUpdateToModel } from '@mail/legacy/component_hooks/use_update_to_model';
 import '@mail/legacy/components/activity_menu_view/activity_menu_view'; // ensure components are registered beforehand.
-import { getMessagingComponent } from "@mail/utils/messaging_component";
+import { getMessagingComponent } from "@mail/legacy/utils/messaging_component";
 
 import { DatePicker } from '@web/core/datepicker/datepicker';
 import { patch } from 'web.utils';
